pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project]
name = "likert_correlation"
version = "0.1.0"
description = "A tool for analyzing correlations in Likert scale data with visualization support"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"

[tool.poetry]
name = "likert_correlation"
version = "0.1.0"
description = "A tool for analyzing correlations in Likert scale data with visualization support"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "likert_correlation"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pandas = "^2.1.0"
numpy = "^1.24.0"
scipy = "^1.11.0"
flask = "^3.0.0"
plotly = "^5.18.0"
openpyxl = "^3.1.0"
pystray = "^0.19.5"
pillow = "^11.1.0"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.9.0"
isort = "^5.12.0"
pyinstaller = "^6.3.0"

[tool.poetry.scripts]
likert-analyzer = "likert_correlation.main:main"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3